fix(INPUT-BUILD-006): add CNA_SANITIZE option; fix signed-overflow UB in input hashes
Add a reusable CNA_SANITIZE CMake option (e.g. -DCNA_SANITIZE=address,undefined)
that instruments CNA + sharp-runtime + tests (vendored SDL/ENet stay prebuilt).

Ran the input filter under ASan+UBSan (cmake-build-input-asan, EasyGL): 279/279
pass, ASan memory-clean. UBSan caught signed-integer overflow in several
GetHashCode implementations -- real UB in C++ (the C# originals relied on
`unchecked` wraparound). Fixed all in-repo occurrences with behavior-preserving
unsigned wraparound (identical 2's-complement result, still matches FNA):

  - Input: GamePadThumbSticks (flagged) + latent GamePadTriggers, GamePadState,
    MouseState, TouchLocation.
  - Shared Framework math: Vector2::GetHashCode (reached via thumbsticks).
  - Updated GamePadThumbSticksTests' expected-hash computation to match (it also
    overflowed).

Hash tests 152/152 still pass (values unchanged); UBSan is now clean for all
in-repo code.

Triaged, NOT fixed here (cross-repo): sharp-runtime TimeSpan copy-ctor is invoked
on an invalid-vptr object during static init of a global DateTimeOffset
(TimeSpan.cpp:54 / DateTimeOffset.cpp:94) -- a sharp-runtime static-init-order
bug, fires at startup regardless of tests. It is the sole remaining UBSan
diagnostic on the input filter and is owned by the sharp-runtime track.
Vector3/Vector4/Color likely share the same hash pattern (Framework follow-up).

+++ src/Microsoft/Xna/Framework/Input/GamePadTriggers.cpp
@@ -49,7 +49,9 @@ namespace Microsoft::Xna::Framework::Input
 
     int GamePadTriggers::GetHashCode() const
     {
-        return System::Single::GetHashCode(left_) + System::Single::GetHashCode(right_);
+        // Unsigned wraparound avoids signed-overflow UB (INPUT-BUILD-006); result is unchanged.
+        return static_cast<int>(static_cast<unsigned>(System::Single::GetHashCode(left_))
+                                + static_cast<unsigned>(System::Single::GetHashCode(right_)));
     }
 
     bool operator==(const GamePadTriggers& left, const GamePadTriggers& right)
